Fix HiveToDruid to allow for non-partitioned source tables.
In the new unique devices druid-loading jobs, the data is prepared
into a temporary table that has no partitioning scheme.
This made HiveToDruid fail, since it was expecting some partitions.
This change allows to load tables that have no partitions.

It also adds a new parameter, count_metric_name, that allows to
change the name of the `count` metric.

Bug: T334096
Change-Id: I4d25b69e920d7f343dbd4df47aac6f43de578b00