Add explicit snapshot to HiveToDruid
The hiveToDuid job can use `since` and `until` parameters to define
an snapshot to load when the snapshot is comprised between the druid
loading interval values.
For snapshot where loading interval doesn't match snapshot interval
(mediawiki_history_reduced or edit_hourly for instance) we create
a new `snapshot` parameter, that when set  overrides the `since`
and `until` parameter for hive snapshot querying.

Change-Id: Ib9af4c7eb66ab8f4213f07b3909e768cb4cdcd48