Add special=ViewObject to allowed pageviews
Wikifunctions uses Special:ViewObject as a proxy for action=view which
avoids going through Varnish.  In this pipeline, this is a weird
exception that will likely need tweaking in the way Varnish handles
X-Analytics to be consistent with other special/non-special page
handling.  But for now, this is the easiest way to get pageviews to
compute properly for wikifunctions and there's not a lot of time left.

Bug: T342865
Change-Id: I6efe490b8a9fd59dcdf3689db421e3809c77a7e4