Retry produceCanaryEvents
Eventgate is sometimes returning some 500 in place of a success code
when POSTing canary events. We may retry the event production.

Example: It returns "eventlogging_VirtualPageView is configured but does not
have any settings". But the stream is configured and well defined.
Likely the next run of the CanaryEvents producer won't trigger an error.
In the meantime, it produces a false alert email.

This patch shouldn't prevent us from finding the root cause.

Bug: T326002
Change-Id: I037369ca7aec26a71837fbb4f052d943fce7ae13