Fix recursion for Maps with Structs on SanitizeTransformation
This patch solves an issue in which the sanitation mask created in
SanitizeTransformation was not able to recurse when the allowlist
includes Maps with struct values.

Additionally, we refactor the tests harness so that tree comparisons
use a tree comparator, for correctness. This was previously alluded
to in 1d76cbc, but it got lost in translation.

Bug: T349121
Change-Id: Ie1435f5bd62dca8ed970b908e2adf3c0730ba25f