Refactor Refine to be triggerd by Airflow
With this changeset:
* We are splitting the Refine mechanism into table management on one side,
and the ETL job to populate the Refined table on the other side.
* And we are doubling the table format options (Iceberg or Hive) pushing
into parameter the transform functions, including normalization. That
means HiveExtension StructType.merge now expect pre-normalized structs.

We are still using Hive JDBC to evolve Hive tables because:
* Spark lack of support for type evolutions
* Incompatibility with Hive once evolved by Spark

The set of supported Iceberg table evolutions:
* add/drop columns/fields
* add/drop/update table properties
* update columns/fields types (widening operations only e.g. int > long)
* update columns/fields comments (drop comments are replaced by a blank string
  override)

The supported Hive DDL operations are unchanged.

Kept for later:
* preserve & evolve fields ordering (sometimes it matters:
  https://phabricator.wikimedia.org/T259924#6374792)
* evolve partitioning

Bug: T356762
Bug: T307040
Change-Id: I8562bb87c89793743bf80f3ed7d22b9bfcc5dc9c