Sort pom.xml according to standard sortpom order.
This is a preliminary step towards migrating to the
wmf-jvm-parent-pom. This includes only resorting the pom with
`./mvnw sortpom:sort` so that the code reviews are simpler
for the next steps.

The order of dependencies are kept as-is for some modules as
there are a number of duplicate classes (jarhell) in this
project and the build fails if the order of dependencies is
changed. A full cleanup of those dependencies should be done
at some point, but ignoring the problem at this step seems simpler.

Bug: T360219
Change-Id: Ib89ce46acce3a651fde025d56fd361af4b5bd145