Start using wmf-jvm-parent-pom.
This standardizes the build process, in particular it introduces
a number of static analysis tools, and fixes versions of Maven
plugins.

Since a number of violations exist currently in this project,
some static analysis have been disabled to ensure the build
passes. Violations will need to be corrected and plugins
re-enabled.

Phab: T360219
Change-Id: I87eec57602998b0362bd230adc2a342af5f85aec