Make TransformFunction deduplication deterministic
Spark deduplication retains the first row of each group of duplicates.
To ensure consistency in the selected row regardless of the process,
we need to sort the Dataset before deduplication.

While refactoring the Refine process, we aim to compare the output of
the new process with the legacy one. This patch will help us do that.

Bug: T356762
Change-Id: I930fd3b11cbdc948b854f309eb298f8e2b46a8a7