Update the smtp server settings for email from refine
The SRE Infrastrcuture Foundations team have made some recent
improvements to the email system configuration and now recommend that
applications use the mail transfer agent (MTA) running on each host,
rather than a remote mail exchanger (MX).

The current MX servers are soon to be decommissioned, so we now need to
update the refinery configuration and remove the configuration that uses
the MX server.

Bug: T325394
Change-Id: Ibfbb1617846b945d176b68d7ba719e506aae8d76