Event deduplication via windowing - backport on 0.2.49
The goal of this patch si to backport 1080306 to the
0.2.49 release. Next release will be named 0.2.49_2.

0.2.49 is only used by legacy Refine triggered from
systemd.

Original path commit msg:

Sorting alone was insufficient to achieve deterministic deduplication.
This patch introduces deduplication using a window function,
leveraging the available sorting column (`meta.dt`).

We now ensure that the most recent event in each duplication group
(`meta.dt desc`) is retained, providing a consistent outcome in the
presence of duplicate events.

Bug: T369845
Change-Id: I5e784c74b0c6feae1aeaf4805016da7c22b1d3de