Backport: Refine transforms minimalist events support to 0.2.49
As defined here,
https://wikitech.wikimedia.org/wiki/Event_Platform/Schemas/Guidelines#Required_fields

a small handfull of columns are required by our
system. We now get an new stream respecting
this spec. But our Refine transform functions
are crashing.

This patch adds a unit test to prevent failures.

This is the backport for branch 0.2.49 used by
legacy Refine.

Original PR is 1112392 .

Bug: T383914
Change-Id: Icc80d116c2281f20608f68fd55eea99af34baf06