Refine transforms minimalist events support
As defined here,
https://wikitech.wikimedia.org/wiki/Event_Platform/Schemas/Guidelines#Required_fields

a small handfull of columns are required by our
system. We now get an new stream respecting
this spec. But our Refine transform functions
are crashing.

This patch adds a unit test to prevent failures.

Note: after review, a new PR needs to be open for
branch 0.2.49 to backport the change to legacy
refine.

Bug: T383914
Change-Id: Ibdaabdabb6a10a50294fa000c97523e6e879ba8b