Allow loading an external regex.yaml file for ua-parser
This will allow us to use a custom regex.yaml file for detection of
user agents, which is useful because we might want to integrate what
is included in the package with some custom regexes based on our traffic.

I made the choice to use an environment variable to set the path to the
regex.yaml file, because that's the easiest way to do it in both a systemd
unit and a docker container.

If the variable is not set, we will just use the default regex.yaml file
that is included in the ua-parser jar.

Bug: T394794
Change-Id: I8b06baf51e9791289f0bed0be27cabb7f31ad4c1