Refine: Fix location of refined data-empty partitions
table location and not baselocation was
sent to create the HivePartition in Refine.
This object is used to add empty-data partitions.

This bug was generating directories for partitions
with an extra table name in its path.

example: hdfs://analytics-hadoop/wmf/data/event/mediawiki_page_create/mediawiki_page_create/datacenter=codfw/year=2025/month=7/day=23/hour=7

In an independant script we will fix the partition
locations for existing misplaced partitions.

This bug does not prevent querying the data.

Bug: T369845
Change-Id: I7e57a6280b9868bc9bf3bfd3314aa79a66cf968d