Add utility to create SHA256 fingerprints of the files of a particular HDFS folder.
This utility reads all files inside the specified HDFS folder, ignoring hidden files and files starting with _.
For all these files, it calculates their SHA256 fingerprint, and writes these fingerprints to a SHA256SUMS file in the
same folder.

This is especially useful to be run after MediaWikiDumper has run, as that process can potentially generate 1000s of
files that a consumer may want to verify after download.

Bug: T401022
Change-Id: I863f812f3b5ed9b230573ebb8978c25fbb97a6a7