Fix bug MW Dumper in which vertical bars ( `|` ) were not being honored.
We compared commonswiki and simplewiki data from legacy Dumps and found out this issue.

The bug was hidden in our use of Scala multi-line Strings and stripMargin(). Thus we take the
opportunity to move XML generation code to use a StringBuilder approach, which,
presumably, should be a bit more performant as well.

Bug: T407649
Change-Id: I4cd56562a23ce03ce819cdd6b305d116dfa67166