Update Refine-CLI job to not overwhelm metastore
The job was using spark dynamic-partition mechanism to overwrite
partitions in hive table, but this mechanism needs to load all table
partitions into the spark catalog, leading to high pressure on the
metastore.
This patch changes the approac using a delete+insert mechanism
preventing the load all partitions into the catalog.

Bug: T410378
Change-Id: I102f1b24fb5fd96c69b34450e7c67e379ebea880